watchdog: Make sure no one can hold a strong reference to the tick counter
The previous code assumed that no one could get hold of a strong reference to
the tick counter in order to determine when to stop the dut_power thread.
This assumption was flawed, as one could simply .upgrade() the Weak Arc
returned by the tick() function.

Make sure no one can hold a strong reference to the tick counter by wrapping it
in a private field of a struct.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>