broker: make retained values more flexible and send retained val on subscribe
This e.g. useful when drawing the diagrams in the web interface where we want
some history right from the start.

Most consumers want to get the most recent value when subscribing and have to
cobble something together using .subscribe_unbounded() / .get() / .try_get().
These are prone to race conditions.
We can just enqueue the most recent value on subscribe by default.

Except when we can not, like in the case of button events.

Now that one can set the maximum retention to 0 values this issue can be worked
around. Not great, not horrible.