dut_power: make OffDischarge the default Off state and OffFloating special
The voltage measurement on the LXA TAC powerboard is quite high impedance.
This is rather nice, as it means we do not load the power supply etc.
What's not nice is that it means the measurement is very susceptible to noise,
e.g. when no DUT is connected.
This has lead to some confusion in the past.

Just make the low-impedance OffDischarge the new "Off" and make OffFloating
the "I know what I am doing" special case.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>