ui/screensaver: Increase Timeout
When working with the LXA TAC and observing some measurement on the
display the short timeout often get's in the way.

This change increases the timeout by a factor of 10.
This should be enough in most cases.
But, compared to lifetime, should still be short enough to prevent
the display from buring in.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>