dut_power: add test for transient resiliance
Short transients on the measured values do not result in damage to the TAC or
DUT, so there is no reason to turn off the power switch.
Add a test for resiliance agains transient values that could appear on the real
hardware due to inrush currents or EMI events.

There is currently no filtering for said events. The test is expected to fail.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>