ci: tacd-webui: fix built errors due to missing autolinker source files
We do not need or use source maps and the build process generates a lot of
warnings due to missing source files in the (indirect) autolinker dependency.
Updating to the most recent autolinker version via npm overrides does not
solve the problem.

Just set the environment variable that disables source map building.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>