web: remove MqttDurationButton and its only user
eslint complains about the useEffect hook missing some dependencies:

  [eslint]
  src/MqttComponents.tsx
    Line 243:6:  React Hook useEffect has missing dependencies:
    'props' and 'setPayload'.
    Either include them or remove the dependency array.
    …

We could work around this warning or properly fix it.
Or we could just remove the MqttDuration button alltogether,
as its usability is quite bad and styling it has also been an unpleasant
endeavour.

Replace the previous use (buttons to control the view of the TACs display)
by two separate buttons for short press and long press.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>