ci: tacd: make cargo deny checks more strict
Cargo deny performs some useful checks on the dependencies of a rust
project.
Some of the issues it points out are however non-trivial to fix,
like in our case duplicate versions of dependencies that are mostly
caused by the tide dependencies that has not seen a release in some
time.

I've gone through the current list of warnings, checked if they affect
use and added them to the ignore list if not.
This allows us to run the checks in the CI and get notified if new issues
crop up.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>