ui: widgets: keep framebuffer locked during clear/draw cycle
Previously the following could happen:

  - Widget clears the dirty area of the screen
  - Framebuffer is drawn to the screen/rendered as png for
    the web interface
  - Widget re-draws itself

This could lead to widgets flickering into and out of existence.
This did not happen (to a noticeable degree) on the actual display,
but did sometimes happen in the virtual view in the web interface.

Keep the draw target locked during the clear/draw cycle to prevent
flickering.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>