adc: make new() fail if iio setup failed
Previously we had problems when the ADC setup failed early (e.g.) because
of the device not being there.
In that case the tacd should just exit or at least be killed by the watchdog,
but that was not the case as the ADC thread just never started producing
values.

Synchronize the IioThread setup with a queue and only continue the tacd
startup once the ADC thread is operational.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>