broker: topic: hand around Values instead of Arc<Values>
Most values we pass around are smaller than a pointer and copying them is
cheaper than doing the atomic reference counting required for Arcs and the
required heap allocations.

Using Arcs makes most sense for values that we want to share a single
instance to to use interior mutability or if the contained data structure
is quite large.
The first applies to no topics at all (as that would make the pub/sub system
pointless) and the second applies to a few.
I've used explicit Arc<>s for topics with large data structures.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>