measurement: add Timestamp type and break out into separate file
Initially only the ADC needed to produce time-stamped measurements,
but then the temperatures module also needed it and just grabbed into
the adc mod to get it.
The actual providers of raw ADC values (iio/{hardware.rs,stub.rs})
used other types again.

Clean up the layering by moving Measurements and Timestamps into their
own module.

In the future we may want to be able to filter retained values based on
their age. For this we may want a Timestamped Trait, which would be easier
to implement with these shared types.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>