adc: add API endpoint to get the current system/adc thread time
This can be used to trigger periodic events in the tacd via the broker
framework (most useful for UI widget animations) but also to give the
web interface some sense of the current time on the TAC.
This can be useful to e.g. warn the user if there is a large discrepancy
between browser time and TAC time, e.g. due to NTP servers being unavailable.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>