adc: iio: hardware: remove obsolete comments
Having Ordering::Acquire for the timestamp loads ensures that all
(even Relaxed) stores to _other_ variables are observed as having
happened before the Ordering::Release store to the timestamp.

This means we do not need fences to synchronize the other variables.

See [1] for examples on when it makes sense to uses a fence.

[1]: https://marabos.nl/atomics/memory-ordering.html#fences

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>