ui: use more idiomatic rust to handle screen change events
The previous implementation comes from a time where the retained values
was not automatically put into the subscribe queue when subscribing.

This is no longer the case. We can now rely on screen_rx.next().await
immediately providing the initial retained value (Screen::ScreenSaver)
after subscribing.

This simplifies the loop.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>