http_server: rename from the somewhat confusing "web"
The web.rs / WebInterface names were a bit more appropriate when the
web interface was actually built into the tacd binary and the code
to do so was located in the web.rs file.

Nowadays it only sets up the http server and some static serves,
most parts of what makes the web interface tick (broker framewor,
typescript code etc.) lives elsewhere.

Rename the module to what it is.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>