dut_power: add OutputState::Changing to prevent losing state transitions
If the DUT power output is off due to an error condition like
OutputState::InvertedPolarity that still persists (e.g. there
is a negative voltage on the output _right now_) the tacd would
previosly not send out a state transition to a client in response
to a change request, as the state would be the same before and
after (OutputState::InvertedPolarity).

The toggle buttons in the web interface have three states they can
be in: On, Off or Unknown. When pressed they change from On or Off
to Unknown and stay grayed out and unklickable until a new known
state is received from the tacd.

Combined this means that the toggle button in the web UI will
be unusable once it was clicked while in an error state until the
side is reloaded.

Prevent that by forcing a state transition on every request, e.g.
OutputState::InvertedPolarity -> OutputState::Changing ->
OutputState::InvertedPolarity.

This should be safer than other fixes I've come up with (timeout on
the slider in the web UI, timeout in the state updating code e.t.c)
with respect to glitchy behaviour.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>