broker: persistence: add option to persist topic values to disk
At first this will only be used to store the "has this TAC been initialized?"
state to the eMMC but later we could also have other persistent settings.
What comes to mind are things like "use OffFloating when requested to turn
the DUT off" or "Use a custom DUT current limit of 1A".

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>