broker: topic: allow setting up anonymous topics from everywhere
There is no reason to require a BrokerBuilder to set up a Topic that is not
intended to be exposed to the outside world.
Just allow setting up anonymous Topics from anywhere and update all call
sites.
This also cleans up some locations that previosly needed a BrokerBuilder
around just to setup an anonymous Topic.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>