broker: topic: use non-async Mutex
Using normal Mutexes in async programs brings with it a risk of panics:

  let shared_res = Mutex::new(0u8);

  async fn task1() {
    let mut res = shared_res.lock().unwrap();

    async_std::task::sleep(Duration::from_secs(10)).await;

    /* implicitly drop the MutexGuard and thus unlock the lock */
  }

  async fn task2() {
    let mut res = shared_res.lock().unwrap();
  }

If task1 runs in a thread, yields and then task2 is scheduled on the same
thread the lock() call in task2 will panic[1], as there would be no way for
the lock to ever become available.

There is one simple way around this: don't carry a MutexGuard across an
.await boundary. We already adhere to that, so we can reap the benefits of
using a normal Mutex instead of an async one:

 - They are much faster
 - We do not litter `async`s absolutely everywhere

The second point is also why this commit touches so many files.

[1]: https://doc.rust-lang.org/1.68.2/std/sync/struct.Mutex.html#method.lock

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>