ui: screens: setup: Add a screen that directs to the setup wizard
Right out of the box the LXA TAC is not yet usable and we have a bit of a
chicken and egg problem to get it fully usable, as new users need a way
to deploy SSH keys on the device, but they do not yet have anything like
SSH keys, that give them administrative access to the device on the device.

This is why we have a privileged setup mode wizard that can deploy SSH keys
(and in the future also trusted RAUC keys) to the device and should afterwards
be disabled.

To be sure that users disable the mode and to make the unboxing experience
more straight-forward we nag them to go to the setup wizard using their web
browser.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>