setup_mode: add a mode in which e.g. the ssh key can be edited from web
We do not want anyone with HTTP access to the TAC to be able deploy SSH keys.
HTTP basic auth over unencrypted connections, maybe even with with default
passwords like "admin:admin" is also kind of pointless.

Go another round (at least for now) of having a priviled setup mode in which
everything is allowed via HTTP and a normal operation mode.
Going back to the setup mode is only possible via local access to the device.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>