ui: display: internalize locking
We generally want to give exclusive ownership to the display to a single
screen at a time.
This does however collide with the need to have a second reference for
screen capturing.
Solve it with an owned Display type that you however still have to lock
to do actual drawing on it.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>