ui: screens: make destruction consume the active screen
Having an unmount(&mut self) method that handles the destruction of a screen
that does not _consume_ the screen (e.g. unmount(self)) is a bit unelegant.

This is the first step towards a design where the exclusive access to
resources like the display can be guaranteed by the type system.

E.g. something like this:

    let screen = ExampleScreen.new();
    let active_screen = screen.activate(display);
    ...
    let display = active_screen.deactivate();

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>