ui: add a concept of alerts
Alerts are screens that have priority over all normal screens and some of
the other alerts.

This means the display will always show the highest priority alert (if one
is active) or one of the normal screens.

Having `NormalScreen`s and `AlertScreen`s separate also allows us to
centralize the handling of navigation for `NormalScreen`s, as they just
transition to the next screen in the list.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>