dbus: rauc: update_channels: add a concept of update channels
On vanilla TACs we are going to have two official update channels
for operating system bundles, Stable and Testing.

In addition to that users may also choose to add their own custom
channels.

Add concept of these channels and a way to poll them to
check if a new update is available.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>