web: replace RaucInstall component with UpdateStatus
While that is not quite decided yet, it is likely that we do not want to
enable installing just any bundle via the web interface (as long as the
signature works out) outside of setup_mode.

Remove the feature from the UI for now and instead rely on update
notifications.

Installation of any bundle for which the certificates work out is however
still possible via the API.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>