http_server: add directory listings for user-provided files
This makes the HTTP serving feature of the TAC more convenient
by giving the user a list of files when requesting a directory without
an index.html inside.

It also makes index.html work in nested directories and prevents an
error message from being shown if the /srv/www directory does not exist
on startup.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>