http_server: serve pre-compressed files if available
If a client requests e.g. index.html with a GET request containing a
"Accept-Encoding: … gzip …" header and an index.html.gz file esists
right next to the original file in the file system which has the same
modification date as the original file serve the compressed file to the
client with a "Content-Encoding: gzip" header.

This reduces the transfered size for e.g. static javascript files by a lot.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>