web: MqttComponents: snap to specified y-axis dimensions
As of now the y-Axis auto-scales to match the values in the plot.
This means it expands the axis to arbitrarily small values, making
every bit of ADC noise look huge.

Specify explicit ranges to snap to instead. The selected values are not
completely random. The DUT voltage ranges for example are choosen just
above typical supply voltages (5V, 12V, 24V, 48V).

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>