adc: iio: hardware: use iio buffering for the powerboard ADC
This means the STM32/baseboard ADC and the powerboard ADC can share
most of their setup routine.
The main difference is that the powerboard ADC uses a software trigger
(iio-trig-hrtimer) instead of a hardware timer like the STM32 ADC does.

The software trigger has to first be created via the configs,
making new_powerboard() a bit different from new_stm32().

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>