rauc: display reboot note based on if other slot would be booted
If you've installed a bundle you likely want to boot it,
regardless of if it is newer or older than the one you are
currently running.

You also do not want to be nagged about rebooting into another slot
if a reboot will only land you in the same slot you are running again.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>