dbus: networkmanager: fix an issue where the IPv4 would not update
The main cause of the issue was that we did not honor the following section
in the NetworkManager DBus API documentation[1]:

  The "Ip4Config" property

  Object path of the Ip4Config object describing the configuration of the
  device. Only valid when the device is in the NM_DEVICE_STATE_ACTIVATED state.

This meant however that we need to keep track of state property changes and
changes inside the referenced Ip4Config object at the same time.
The easiest way to do so was to re-write the code in a async/.await-y way.

[1]: https://networkmanager.dev/docs/api/latest/gdbus-org.freedesktop.NetworkManager.Device.html#gdbus-property-org-freedesktop-NetworkManager-Device.Ip4Config

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>