ui: reduce idle brightness of the status led
The first hardware samples of LXA TACs with status LEDs showed them
to be far to bright.
Since then we have reduced the current through the "normal" (non-pwm)
status LEDs by a factor of 33 in hardware.
The brightness of the pwm LEDs was only reduced in hardware by a factor
of 7.5. This allows us to still make them light up quite brightly if
we want them to.

In the idle state their brightness should however match the brightness
of the other LEDs.

Dim them in software by a factor of 33/7.5 = 4.4.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>