web: TacComponents: UsbOverloadNotification: fix note on disconnect
The topic value can both be `null` (no overload error was set by the
host) and `undefined` (no value was set yet or the host is currently
disconnected).
Check for both when deciding whether to show the notification or not
to prevent it from being from shown on connection loss.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>