digital_io: add clippy exception for the #[cfg(test)] variant of mod gpio
Clippy disallows #[cfg(test)] mods that are followed by other code by
default[1] and usually this makes sense, because tests should be at the
end of the file, but in our case the mod does not contain tests but instead
imports a test-specific variant of the gpio mod.

Allow that.

[1]: https://rust-lang.github.io/rust-clippy/master/index.html#/items_after_test_module

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>