rauc: wait two minutes after startup before polling for updates
The first poll for updates after a fresh boot usually fails,
because the network is not yet up.
This means an update notification will only appear once one
polling interval passed since startup (which can be on the
order of days).
We could use a heuristic to decide once the network is up
so we can start polling, but we are in no hurry and can just
wait a bit before the first poll.
"A bit" being two minutes in this case because it feels about.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>