ui: screens: make draw_border take a DisplayExclusive instead of Display
Currently draw_border does its own locking of the display,
but most callees also lock the screen to draw UI elements or
will do so soon when button legends are introduced.

Make draw_border use the already locked display instead of doing
its own locking.

This is mostly an esthetic change, as the framebuffer backend is
not at all concerned with the locking we do here and performance-wise
it should not make a difference.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>