ui: add button labels that tell the user which button does what
The legend is drawn at the right of the screen using rotated text.
The embedded_graphics library does not have support for rotating
elements like text, so we instead provide it with a _view_ of the
screen that is rotated and onto which it can render its text as
normal.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>