dut_power: implement an error-free grace period after DUT power on
We have seen errors where the DUT power switch refuses to turn on when
a power supply is connected but the OUT- pin of the power switch is not
connected to the GND level of the TAC.

The voltage measurements that lead to this decision are actually correct,
because the voltage measurement is very high impedance and small leakage
currents make their way from the power supply into the power switch.

But being correct and being sensible are two different things.

Make it so that one can always turn the output on, leave it on for a bit
to let everything settle and only then start to check for errors on the
voltage supply.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>