web: LabgridConfig: Render labgrid yaml config files as text
labgrid's yaml configuration files can contain Jinja2 for templating.
This is, for example, used in `/etc/labgrid/configuration.yaml` that is
shipped with `meta-lxatac`.

Since the `yaml` syntax highlighter of the editor is not able to handle
this, it displays syntax errors. Even if the actual file is totally
correct.

This behavior already led to users needing support while initially
setting up their devices.

This change disables the `yaml` syntax highlighter.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>