web: setup: Inform user about data sent when polling for updates
When polling for updates we want to transmit some information about the
device back to our servers.
(This is intended for us to better understand how the updates progress in
a fleet of devices.)

We want to allow the user to make an informed decision when enabling
updates, if they want to give consent to this behavior.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>