ui: help: Start first screen with a call to long-press
Showing users the help (aka quick-start) screen has shown that it's not
clear to everyone that they have to long-press the lower button to
actually switch to the next page of the help.

This change adds an explicit call-to-action to long-press the lower
button to the first page of help.
And afterwards explains what this has done.
This makes the help one page longer - but hopefully helps users discover
the long-press feature a little quicker.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>