broker: persistence: downgrade unknown keys in state file from err to warn
This makes it so that a tacd will not fail to start if it encounters
a state file that was written by a newer tacd version.

This means we actually have to use the versioning scheme we already have
when a new persistence key e.g. has security implications that may not
be lost in e.g. a new -> old -> new cycle.

Once such a change occurs we have the option to:

  - Increment the `format_version` field in the state file and make older
    tacds fail to start that way.
  - Start writing changes to a new state file `/srv/tacd/state-v2.json`
    and also read from there if it exists, but fall back to the current
    path if it does not.
    This means old tacds keep reading the file they created and a new
    tacd will read from there and migrate.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>