adc: determine ADC lines based on hardware generation
The ADC channels 0 and 1 on the STM32MP1 are a bit special,
in that they are not muxable as GPIOs but connected directly to the
ADC peripheral.
This has cause us a bit of trouble in the past, because these pins
seem to leak current (like all other channels did until the transition
into ADC standby was fixed by unmuxing the ADC pins beforehand).
We could once again try to fix the issue with these pins,
but that could also turn out to be impossible to do (since the pins
can not be muxed away).
This is why we have instead decided to just not use these pins on
generation 3 (and likely later) hardware and use other pins instead.

This means we have to detect which ADC pins we should use based on
if we are running on Gen 1 / Gen 2 or Gen 3.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>