ci: tacd: drop rust version 1.68 from mickledore, add 1.75 from scarthgap
The current meta-lxatac / tacos stable release (2024.04) is based on yocto
nanbield (which ships rust 1.70) we need to stay compatible with that to
be able to produce potential point releases.

We no longer need to support the rust version from the yocto mickledore
release (1.68), so we can drop it from the CI tests.

Development will continue based on yocto scarthgap, so we need to add the
rust version it ships (1.75) in our CI tests.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>