dut_power: make turn_off_with_reason() fallible
The operating system call to set the GPIO output state may fail.
Make sure that this results in an orderly teardown of the tacd instead
of a panic.

This also necessitates a change to the test and demo_mode gpio stubs,
so they can be `?`-operatored into anyhow Results (simply by making
them anyhow results).

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>