system: also read out the baseboard and powerboard featuresets
This adds support to read the featurset fields stored in the baseboard and
powerboard EEPROMs which are passed to us by barebox via the devicetree
choosen mechanism.

The featuresets describe things like "does this TAC have a TFT"
(in contrast to the OLEDs the first generation of TACs had)
or "are these scales and offsets actually calibrated or only
calculated".

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>