ci: tacd: update to cargo deny version 2 format
The new format removes the unlicensed and default keys
(making deny the default for them and providing other options to override
it - which we do not need) and removes the copyleft field which allowed
all GPL-style licenses at once.
Instead we now have to list the LPGL (for systemd) and GPL-2.0 (for tacd)
ourself.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>