web: remove unused App.test.tsx
The App.test.tsx file was created by create-react-app when initially setting
up the project. It has not been updated or used since then though.

If the parameters of the `App` function ever where to change (foreshadowing)
the build would break, requiring changes to App.test.tsx.

Since the file is unused it can just be removed instead.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>