web: legal: test: prepare everything for running tests with jest
The create-react-app project has already prepared everything for us so we
can run unit tests on our code base.
All that is left to set up are some tweaks so that jest runs babel on the
cloudscape dependency, because otherwise the tests will trip over
javascript features that are just too new.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>