web: downgrade dependencies by checking out an old version
The upgraded dependencies have exposed an issue in the way package names
are handled in the bitbake npm package fetcher.
The issue being that the names '@types/ramda' and 'types-ramda' are
mangled into the same name, which results in their tar archives being
called the same if the versions match as well,
resulting in one overwriting the other.
There is a fix for this, which made it into poky master[1] but not into
scarthgap (yet).

Getting it into scarthgap is the ultimate goal, but right now it is
holding us back from rolling a new release.

Undo the the dependency updates for now, so we can roll a new meta-lxatac
release. We can re-do the updates once the fix is backported to scarthgap.

This resets the package-lock.json to how it was before commit cee818c
("tacd: update dependencies via cargo update").
In this package-lock.json the '@types/ramda' and 'types-ramda' packages
have different version numbers, which means the tar archives have
different names.

[1]: https://github.com/yoctoproject/poky/commit/ee6bf285d7ca35ab0f76c151a36e888ad1e41c40

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>