ci: tacd: use run: cargo deny instead of EmbarkStudios/cargo-deny-action
Other than the actions-rs actions the cargo-deny action is actually
actively maintained, but replacing it with a simple `cargo deny` call
removes our last non-GitHub provided action dependency and makes it a bit
easier to determine what is actually happening in the action,
should one e.g. want to reproduce it locally.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>