ci: tacd-webui: use a fixed node version instead of latest
The build sometimes breaks due to node updates.
This would not be much of a problem if the fixes for these breakages did
not break our actual builds with the node versions used in Yocto.

Fix that by using a fixed node version - in this case the one from the
Styhead release.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>