cargo deny: acknowledge that "surf" is unmaintained
The advisory[1] says the following about surf:

    The last release is over three years old (from 2021), the crate
    depends on the deprecated `async-std` crate and on a very old version
    of `rustls` for TLS support.

We use `surf` _because_ it is `async-std` based and we use `async-std`
throughout our code base, which means we will throw it out once we bite
the bullet and replace `async-std`.

We only use surf to poll the local `lxa-iobus` REST API, so we do not have
to fear possible security issue so much.

[1]: https://rustsec.org/advisories/RUSTSEC-2025-0036

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>